core/state/snapshot, cmd/geth: polish code (#37)
